tests: Enable macro disabling
This commit introduces new cmake option "WITH_DLT_DISABLE_MACRO" to
disable macro usage in case it is not allowed. Currently this is adapted
to src/tests and src/examples only. for dlt-test-non-verbose, the build itself is
currently disabled if WITH_DLT_DISABLE_MACRO is set; this needs to be
enabled again in future by replacing or adding function usage.

Signed-off-by: Saya Sugiura <ssugiura@jp.adit-jv.com>