daemon: Enable to use FIFO on QNX
It depends on OSes if maximum size of pipe can be configured per
process or not. For example, on QNX it can be only set from pipe
resource manager. Also flags used to set the value are only supported
on Linux.
Now DaemonFIFOSize option in dlt.conf is only relevant for Linux.

Also -f option is added to build scripts to enable FIFO as IPC.

Signed-off-by: Saya Sugiura <ssugiura@jp.adit-jv.com>