debian: improve debian build package
Currently, when building dlt debian package, our libdlt.so can not be linked to
other binary consoles.
Adding build option of dh_shlibdeps to specific the path of libdlt.so, and
linking step should not be broken.

With default enabling flag WITH_MAN in debian rules, pandoc package must be
added in Build_Depends option for debian control.

Signed-off-by: Dinh Cong Toan <toan.dinhcong@vn.bosch.com>